Move `parse_points` to panel UI
