+++ audiovlm_demo/core/components.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 import gc
 import io
-import re
 import time
 from pathlib import Path
 from typing import Annotated, Any
@@ -152,30 +151,6 @@ class AudioVLM:
                 self.model_store["Loaded"] = True
             case _:
                 pass
-
-    @classmethod
-    def parse_points(cls, points_str: str):
-        # Regex to extract each <points> tag with multiple x and y pairs
-        point_tags = re.findall(r"<points (.*?)>(.*?)</points>", points_str)
-        if len(point_tags) == 0:
-            point_tags = re.findall(r"<point (.*?)>(.*?)</point>", points_str)
-        parsed_points = []
-        if len(point_tags) == 0:
-            return None
-
-        for attributes, label in point_tags:
-            coordinates = re.findall(r'x\d+="(.*?)" y\d+="(.*?)"', attributes)
-            if not coordinates:
-                single_coordinate = re.findall(r'x="(.*?)" y="(.*?)"', attributes)
-                if single_coordinate:
-                    coordinates = [single_coordinate[0]]
-            parsed_points.append(
-                {
-                    "label": label,
-                    "coordinates": [(float(x), float(y)) for x, y in coordinates],
-                }
-            )
-        return parsed_points
 
     _default_system_prompt = "You are an unbiased, helpful assistant."
